pyproject.toml
[project]
name = "cognee-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cognee>=0.4.1",
    "playwright>=1.56.0",
    "protego>=0.5.0",
    "python-dotenv>=1.2.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
]

[dependency-groups]
dev = [
    "ruff>=0.14.7",
    "ty>=0.0.1a29",
]
